Revert lock implementation in Store
