+++ VueFlux/ThreadSafe.swift
@@ -1,7 +1,14 @@
+import Foundation
+
 /// Thread-safe value wrapper.
 public final class ThreadSafe<Value> {
     private var _value: Value
-    private let lock: Lock
+    private let lock: NSLocking = {
+        if #available(*, iOS 10.0, macOS 10.12, tvOS 10.0, watchOS 3.0) {
+            return OSUnfairLock()
+        }
+        return PosixThreadMutex()
+    }()
     
     /// Synchronized value getter and setter
     public var value: Value {
@@ -13,10 +20,8 @@ public final class ThreadSafe<Value> {
     ///
     /// - Parameters:
     ///   - value: Initial value.
-    ///   - recursive: A Bool value indicating whether lock is recursive.
-    public init(_ value: Value, recursive: Bool = false) {
+    public init(_ value: Value) {
         _value = value
-        lock = .initialize(recursive: recursive)
     }
     
     /// Perform a given action with current value thread-safely.
@@ -27,9 +32,9 @@ public final class ThreadSafe<Value> {
     /// - Returns: Result value of action.
     @discardableResult
     public func synchronized<Result>(_ function: (Value) throws -> Result) rethrows -> Result {
-        return try lock.synchronized {
-            try function(_value)
-        }
+        lock.lock()
+        defer { lock.unlock() }
+        return try function(_value)
     }
     
     /// Modifies the value thread-safely.
@@ -40,9 +45,9 @@ public final class ThreadSafe<Value> {
     /// - Returns: Result value of modification action.
     @discardableResult
     public func modify<Result>(_ function: (inout Value) throws -> Result) rethrows -> Result {
-        return try lock.synchronized {
-            try function(&_value)
-        }
+        lock.lock()
+        defer { lock.unlock() }
+        return try function(&_value)
     }
     
     /// Set the new value and returns old value.
@@ -57,6 +62,62 @@ public final class ThreadSafe<Value> {
             let oldValue = value
             value = newValue
             return oldValue
+        }
+    }
+}
+
+private extension ThreadSafe {
+    @available(iOS 10.0, *)
+    @available(macOS 10.12, *)
+    @available(tvOS 10.0, *)
+    @available(watchOS 3.0, *)
+    private final class OSUnfairLock: NSLocking {
+        private let _lock = os_unfair_lock_t.allocate(capacity: 1)
+        
+        init() {
+            _lock.initialize(to: os_unfair_lock())
+        }
+        
+        deinit {
+            _lock.deinitialize()
+            _lock.deallocate(capacity: 1)
+        }
+        
+        func lock() {
+            os_unfair_lock_lock(_lock)
+        }
+        
+        func unlock() {
+            os_unfair_lock_unlock(_lock)
+        }
+    }
+    
+    private final class PosixThreadMutex: NSLocking {
+        private let _lock = UnsafeMutablePointer<pthread_mutex_t>.allocate(capacity: 1)
+        
+        init() {
+            _lock.initialize(to: pthread_mutex_t())
+            
+            let result = pthread_mutex_init(_lock, nil)
+            assert(result == 0)
+        }
+        
+        deinit {
+            let result = pthread_mutex_destroy(_lock)
+            assert(result == 0)
+            
+            _lock.deinitialize()
+            _lock.deallocate(capacity: 1)
+        }
+        
+        func lock() {
+            let result = pthread_mutex_lock(_lock)
+            assert(result == 0)
+        }
+        
+        func unlock() {
+            let result = pthread_mutex_unlock(_lock)
+            assert(result == 0)
         }
     }
 }
